Include the fff library for mocking functions.
So far, I like this library. However, it may need to be tweaked to play nicely
with wrapped functions, which is the way we override functions in the xinu
linker:

In order to override a function, say, `netin`, we must defined a function
`__wrapped_netin`. The linker is configured to detect this and link it into
place. Inconveniently, though, that means we have to refer to `__wrapped_netin`
when the mock is defined, and to mock macros.

I've tried to ease this by putting `#define netin __wrapped_in` at the top of a
test, but this is unwieldy. Still thinking of a potential better way to do
this.